+++ src/RGBAColor/ColorChannel/AlphaChannel.php
@@ -2,7 +2,7 @@
 /**
  * @package Application Utils
  * @subpackage RGBAColor
- * @see \AppUtils\RGBAColor\ColorChannel\PercentChannel
+ * @see \AppUtils\RGBAColor\ColorChannel\AlphaChannel
  */
 
 declare(strict_types=1);
@@ -13,38 +13,35 @@ use AppUtils\RGBAColor\ColorChannel;
 use AppUtils\RGBAColor\UnitsConverter;
 
 /**
- * Color channel with values from 0 to 100.
+ * Color channel with values from 0.0 to +1.0.
  *
- * Native value: {@see self::getPercent()} and
- * {@see self::getPercentRounded()}.
+ * Native value: {@see self::getAlpha()}.
  *
  * @package Application Utils
  * @subpackage RGBAColor
  * @author Sebastian Mordziol <s.mordziol@mistralys.eu>
  */
-class PercentChannel extends ColorChannel
+class AlphaChannel extends ColorChannel
 {
     public const VALUE_MIN = 0.0;
-    public const VALUE_MAX = 100.0;
+    public const VALUE_MAX = 1.0;
 
     /**
      * @var float
      */
-    protected float $value;
+    private float $value;
 
-    /**
-     * @param int|float $value
-     */
-    public function __construct($value)
+    public function __construct(float $value)
     {
-        $value = (float)$value;
-
         if($value < self::VALUE_MIN) { $value = self::VALUE_MIN; }
         if($value > self::VALUE_MAX) { $value = self::VALUE_MAX; }
 
         $this->value = $value;
     }
 
+    /**
+     * @return float 0.0-1.0
+     */
     public function getValue() : float
     {
         return $this->value;
@@ -52,26 +49,26 @@ class PercentChannel extends ColorChannel
 
     public function getAlpha() : float
     {
-        return UnitsConverter::percent2Alpha($this->value);
+        return $this->value;
     }
 
     public function get8Bit() : int
     {
-        return UnitsConverter::percent2IntEightBit($this->value);
+        return UnitsConverter::alpha2IntEightBit($this->value);
     }
 
     public function get7Bit() : int
     {
-        return UnitsConverter::percent2IntSevenBit($this->value);
+        return UnitsConverter::alpha2IntSevenBit($this->value);
     }
 
     public function getPercent() : float
     {
-        return $this->value;
+        return UnitsConverter::alpha2percent($this->value);
     }
 
-    public function invert() : PercentChannel
+    public function invert() : AlphaChannel
     {
-        return ColorChannel::percent(100-$this->value);
+        return ColorChannel::alpha(1-$this->value);
     }
 }