+++ src/RGBAColor/ColorChannel/EightBitChannel.php
@@ -1,4 +1,9 @@
 <?php
+/**
+ * @package Application Utils
+ * @subpackage RGBAColor
+ * @see \AppUtils\RGBAColor\ColorChannel\EightBitChannel
+ */
 
 declare(strict_types=1);
 
@@ -7,16 +12,40 @@ namespace AppUtils\RGBAColor\ColorChannel;
 use AppUtils\RGBAColor\ColorChannel;
 use AppUtils\RGBAColor\UnitsConverter;
 
+/**
+ * Color channel with values from 0 to 255.
+ *
+ * Native value: {@see self::get8Bit()} and
+ * {@see self::getHexadecimal()}.
+ *
+ * @package Application Utils
+ * @subpackage RGBAColor
+ * @author Sebastian Mordziol <s.mordziol@mistralys.eu>
+ */
 class EightBitChannel extends ColorChannel
 {
+    public const VALUE_MIN = 0;
+    public const VALUE_MAX = 255;
+
     /**
      * @var int
      */
-    private $value;
+    private int $value;
 
     public function __construct(int $value)
     {
+        if($value < self::VALUE_MIN) { $value = self::VALUE_MIN; }
+        if($value > self::VALUE_MAX) { $value = self::VALUE_MAX; }
+
         $this->value = $value;
+    }
+
+    /**
+     * @return int 0-255
+     */
+    public function getValue() : int
+    {
+        return $this->value;
     }
 
     public function get8Bit() : int
@@ -29,7 +58,7 @@ class EightBitChannel extends ColorChannel
         return UnitsConverter::intEightBit2IntSevenBit($this->value);
     }
 
-    public function getDecimal() : float
+    public function getAlpha() : float
     {
         return UnitsConverter::intEightBit2Float($this->value);
     }